requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
pyaes==1.6.1
pyasn1==0.4.8
pymongo==3.10.1
pyTelegramBotAPI==3.7.1
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
Telethon==1.14.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
